make _read_adc() unsigned

+++ modules/thermo-cycler/thermo-cycler-arduino/thermistorsadc.cpp
@@ -96,7 +96,7 @@ float ThermistorsADC::heat_sink_temperature() {
 }
 
 
-int ThermistorsADC::_read_adc(int index) {
+uint16_t ThermistorsADC::_read_adc(int index) {
   int res = 0;
   switch (int(index / ADC_PER_DEVICE)) {
     case 0:
@@ -108,12 +108,12 @@ int ThermistorsADC::_read_adc(int index) {
   }
   if (res >=0)
   {
-    return res;
+    return uint16_t(res);
   }
-  return 0;
+  return uint16_t(0);
 }
 
-float ThermistorsADC::_adc_to_celsius(int _adc) {
+float ThermistorsADC::_adc_to_celsius(uint16_t _adc) {
   if (_adc < TABLE[ADC_TABLE_SIZE-1].adc_reading) {
     return TABLE[ADC_TABLE_SIZE-1].celsius;
   }
